Added alternative CSV parsing
In the first version: for every event, the plugin would go through every CSV row, regex match all the IP's in the IP column, then try to match the event IP with the matched IP's.

In this version: when the CSV is loaded (and when it's reloaded every 5 minutes), every row that contains IP's in the IP column is broken up into multiple rows, basically adding redundant information, but simplifying the process of IP matching for every event.